+++ CHANGELOG.md
@@ -3,9 +3,12 @@ CHANGELOG
 [3.0.0]
 
 - .NET 9 suppport
-- **BREAKING CHANGE**: ToGuid conversion done in native endianess (always big endian before)
-  - this was to match `Guid.CreateVersion7()` behavior by default
-  - old behavior possible by using `ToGuid(matchGuidEndianness: false)`
+- **BREAKING CHANGE**: Guid conversion is done in native endianess (was always big endian before)
+  - this was done to match `Guid.CreateVersion7()` behavior by default
+  - old behavior possible by using `ToGuid(bigEndian: true)`
+- **BREAKING CHANGE**: all overloads that used `matchGuidEndianess` were changed to `bigEndian` parameter.
+  - this was done to match most of .NET 9 overloads
+  - old behavior possible by using negation, e.g., `ToGuid(bigEndian: !matchGuidEndiness)`
 - removed obsoleted functions
 - removed official .NET 6 support (code is left behind, just not tested)